+++ include/flexflow/ops/dropout.h
@@ -1,17 +1,12 @@
 #ifndef _FLEXFLOW_DROPOUT_H
 #define _FLEXFLOW_DROPOUT_H
 
-#include "flexflow/device.h"
-#include "flexflow/fftype.h"
 #include "flexflow/layer.h"
 #include "flexflow/node.h"
-#include "flexflow/op_meta.h"
 #include "flexflow/operator.h"
 #include "flexflow/ops/dropout_params.h"
 
 namespace FlexFlow {
-
-class DropoutMeta;
 
 class Dropout : public Op {
 public:
@@ -50,20 +45,6 @@ public:
                             std::vector<Legion::PhysicalRegion> const &regions,
                             Legion::Context ctx,
                             Legion::Runtime *runtime);
-  static void forward_kernel(DropoutMeta *m,
-                             float const *input_ptr,
-                             float *output_ptr,
-                             ffStream_t stream);
-  static void forward_kernel_wrapper(DropoutMeta *m,
-                                     float const *input_ptr,
-                                     float *output_ptr);
-  static void backward_kernel(DropoutMeta *m,
-                              float const *output_grad_ptr,
-                              float *input_grad_ptr,
-                              ffStream_t stream);
-  static void backward_kernel_wrapper(DropoutMeta *m,
-                                      float const *output_grad_ptr,
-                                      float *input_grad_ptr);
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
@@ -79,25 +60,6 @@ public:
 public:
   float rate;
   unsigned long long seed;
-};
-
-class DropoutMeta : public OpMeta {
-public:
-  DropoutMeta(FFHandler handle,
-              Dropout const *dropout,
-              Legion::Memory gpu_mem,
-              Legion::Domain const &output_domain);
-  ~DropoutMeta(void);
-  Realm::RegionInstance reserveInst;
-#if defined(FF_USE_CUDA) || defined(FF_USE_HIP_CUDA)
-  cudnnTensorDescriptor_t inputTensor, outputTensor;
-  cudnnDropoutDescriptor_t dropoutDesc;
-#else
-  miopenTensorDescriptor_t inputTensor, outputTensor;
-  miopenDropoutDescriptor_t dropoutDesc;
-#endif
-  void *reserveSpace, *dropoutStates;
-  size_t reserveSpaceSize, dropoutStateSize;
 };
 
 }; // namespace FlexFlow
