+++ src/ops/kernels/element_binary_kernels.cu
@@ -268,21 +268,21 @@ void forward_kernel(ElementBinaryMeta const *m,
     // currently only handle add and sub
     assert(m->op_type == OP_EW_SUB || m->op_type == OP_EW_ADD ||
            m->op_type == OP_EW_MUL);
-    if(m->op_type == OP_EW_SUB || m->op_type == OP_EW_ADD){
+    if (m->op_type == OP_EW_SUB || m->op_type == OP_EW_ADD) {
       // output = (beta*output + alpha1*input1) + beta*output = input1
       checkCUDNN(cudnnOpTensor(m->handle.dnn,
-                              m->opDesc,
-                              &beta,
-                              m->outputTensor,
-                              out_ptr,
-                              &alpha1,
-                              m->input1Tensor,
-                              in1_ptr,
-                              &beta,
-                              m->outputTensor,
-                              out_ptr));
-      // output = (beta*output + alpha2*input2) + alpha1*output = alpha2*input2 
-      // + alpha1*input1 
+                               m->opDesc,
+                               &beta,
+                               m->outputTensor,
+                               out_ptr,
+                               &alpha1,
+                               m->input1Tensor,
+                               in1_ptr,
+                               &beta,
+                               m->outputTensor,
+                               out_ptr));
+      // output = (beta*output + alpha2*input2) + alpha1*output = alpha2*input2
+      // + alpha1*input1
       checkCUDNN(cudnnOpTensor(m->handle.dnn,
                                m->opDesc,
                                &beta,
@@ -294,7 +294,7 @@ void forward_kernel(ElementBinaryMeta const *m,
                                &alpha1,
                                m->outputTensor,
                                out_ptr));
-    } else if(m->op_type == OP_EW_MUL) {
+    } else if (m->op_type == OP_EW_MUL) {
       checkCUDNN(cudnnSetOpTensorDescriptor(m->opDesc,
                                             CUDNN_OP_TENSOR_ADD,
                                             CUDNN_DATA_FLOAT,
@@ -437,16 +437,16 @@ void backward_kernel(ElementBinaryMeta const *m,
             in1_grad_ptr));
       } else {
         checkCUDNN(cudnnOpTensor(m->handle.dnn,
-                                m->opDesc,
-                                &alpha1,
-                                m->outputTensor,
-                                out_grad_ptr,
-                                &alpha2,
-                                m->input2Tensor,
-                                in2_ptr,
-                                &beta,
-                                m->input1Tensor,
-                                in1_grad_ptr));
+                                 m->opDesc,
+                                 &alpha1,
+                                 m->outputTensor,
+                                 out_grad_ptr,
+                                 &alpha2,
+                                 m->input2Tensor,
+                                 in2_ptr,
+                                 &beta,
+                                 m->input1Tensor,
+                                 in1_grad_ptr));
       }
     }
     if (in2_grad_ptr != nullptr) {
@@ -477,16 +477,16 @@ void backward_kernel(ElementBinaryMeta const *m,
             in2_grad_ptr));
       } else {
         checkCUDNN(cudnnOpTensor(m->handle.dnn,
-                                m->opDesc,
-                                &alpha1,
-                                m->outputTensor,
-                                out_grad_ptr,
-                                &alpha2,
-                                m->input1Tensor,
-                                in1_ptr,
-                                &beta,
-                                m->input2Tensor,
-                                in2_grad_ptr));
+                                 m->opDesc,
+                                 &alpha1,
+                                 m->outputTensor,
+                                 out_grad_ptr,
+                                 &alpha2,
+                                 m->input1Tensor,
+                                 in1_ptr,
+                                 &beta,
+                                 m->input2Tensor,
+                                 in2_grad_ptr));
       }
     }
   } else {
